Genex: Don't segfault on transitive properties in link implementation.
Commit v2.8.11~156^2~2 (Expand includes and defines transitively
in 'external' genexes., 2013-02-13) introduced a recursive loop
and a segfault during evaluation of a link implementation which
depends on a transitive property, such as

 add_library(empty1 ...)
 add_library(empty2 ...)
 target_link_libraries(empty1
   PRIVATE
     $<$<STREQUAL:$<TARGET_PROPERTY:INCLUDE_DIRECTORIES>,/foo/bar>:empty2>
 )

There is no use-case for code like that currently, but it should not
segfault.

Avoid the recursion by reporting an error early if a case like this
is found.

Add a backtrace to link interface access APIs and use it while
transitively evaluating generator expressions through the link
interface. This allows reporting the target_link_libraries line
which specifies a target with a problematic INTERFACE property.