CMP0022: Fix link language propagation in NEW behavior
The languages used in compiling static libraries need to be propagated
to dependents unconditionally.  Factor the logic out of the main link
interface conditions so that it occurs when CMP0022 is NEW and there are
no explicit interface link libraries specified.

Update Tests/Fortran to set CMP0022 to NEW to test this case.