add support for ThreadSanitizer
This commit adds support for ThreadSanitizer to ctest.  ThreadSanitizer
is part of the clang compiler and also gcc 4.8 and later. You have to
compile the code with special flags. Then your code gets the the
ThreadSanitizer ability built into it. To pass options to the ThreadSanitizer
you use an environment variable. This commit teaches ctest to parse
the output from ThreadSanitizer and send it to CDash.